+++ tests/test_routing_v2.py
@@ -27,7 +27,7 @@ def test_default_auto_allow_blocks_explicit_only_providers():
     assert auto_allow["octen"] is False
     assert auto_allow.get("brave", True) is True
     assert set(auto_allow) == {
-        "serpbase", "querit", "donsetch", "octen", "tinyfish",
+        "serpbase", "querit", "donsetch", "octen", "search1api", "tinyfish",
     }
 
 
@@ -43,7 +43,7 @@ def test_legacy_auto_allow_config_inherits_new_guarded_provider_defaults():
     assert validated["auto_routing"]["auto_allow"]["donsetch"] is False
     assert validated["auto_routing"]["auto_allow"]["octen"] is False
     assert set(validated["auto_routing"]["auto_allow"]) == {
-        "serpbase", "querit", "donsetch", "octen", "tinyfish",
+        "serpbase", "querit", "donsetch", "octen", "search1api", "tinyfish",
     }